refactor(Table): rename tableStory.testutils to stories.testutils and update imports
